+++ src/i18n/locales/index.js
@@ -1,0 +1,7 @@
+import ptBrTranslations from './pt-pt'
+import enUsTranslations from './en-us'
+
+export default {
+  'pt-PT': ptBrTranslations, 
+  'en-US': enUsTranslations
+}
