+++ src/i18n/locales/pt-pt.js
@@ -1,0 +1,121 @@
+import { peoplerandom, eco, speed, gear, origamiVase, rasenVase, simpleVase } from "../assets";
+
+export const features = [
+  {
+    id: "feature-1",
+    icon: gear,
+    title: "Personalização Exclusiva",
+    content:
+      "Dê vida às suas ideias com a nossa avançada tecnologia de impressão 3D, oferecendo personalização incomparável. Criamos produtos únicos e específicos para cada cliente, que são tão funcionais quanto inovadores.",
+  },
+  {
+    id: "feature-2",
+    icon: speed,
+    title: "Prototipagem Rapida",
+    content:
+      "Acelere o seu processo de design com os nossos serviços de prototipagem rápida. Entregamos protótipos de alta qualidade de forma rápida, ajudando-o a reduzir o tempo de lançamento no mercado e a manter-se à frente da concorrência.",
+  },
+  {
+    id: "feature-3",
+    icon: eco,
+    title: "Produção Ecológica",
+    content:
+      "Comprometa-se com a sustentabilidade com a nossa impressão 3D ecologicamente consciente. Minimizamos o desperdício e utilizamos materiais recicláveis, garantindo que os seus projetos sejam inovadores e amigos do ambiente.",
+  },
+];
+
+
+export const feedback = [
+  {
+    id: "feedback-1",
+    content:
+      "Uma experiência agradável e profissional. Não podíamos estar mais satisfeitos com os produtos finais. Altamente recomendado!",
+    name: "Margarida M.",
+    title: "People Opps Specialist",
+    img: peoplerandom,
+  },
+  {
+    id: "feedback-2",
+    content:
+      "A equipa foi incrivelmente prestável durante todo o processo, e o desconto de 20% na terceira encomenda valeu a pena.",
+    name: "Luis R.",
+    title: "Engenharia Mecânica",
+    img: peoplerandom,
+  },
+  {
+    id: "feedback-3",
+    content:
+      "A equipa foi muito simpática comigo. Os modelos ficaram ótimos. Recomendo se estiver à procura de algo único!",
+    name: "Maria N.",
+    title: "Professora",
+    img: peoplerandom,
+  },
+];
+
+export const products = [
+  {
+    id: "product-1",
+    content:
+      "Um vaso simples feito para as suas flores.",
+    name: "Vaso Simples",
+    sizes: "L, M, S",
+    img: simpleVase,
+  },
+  {
+    id: "product-2",
+    content:
+      "Dê um toque de modernidade à sua cozinha.",
+    name: "Vaso Rasen",
+    sizes: "L, M, S",
+    img: rasenVase,
+  },
+  {
+    id: "product-3",
+    content:
+      "Um vaso único com o toque certo de modernidade.",
+    name: "Vaso Origami",
+    sizes: "L, M, S",
+    img: origamiVase,
+  },
+];
+
+export const stats = [
+  {
+    id: "stats-1",
+    title: "Clientes Atendidos",
+    value: "100+",
+  },
+  {
+    id: "stats-2",
+    title: "Modelos Pre-feitos",
+    value: "20+",
+  },
+  {
+    id: "stats-3",
+    title: "Dias de Envio",
+    value: "5",
+  },
+];
+
+export const footerLinks = [
+  {
+    title: "Useful Links",
+    links: [
+      {
+        id:"footer-link-1",
+        name: "Termos & Condições",
+        link: "terms",
+      },
+      {
+        id:"footer-link-2",
+        name: "Política de Privacicade",
+        link: "privacy-policy",
+      },
+      {
+        id:"footer-link-3",
+        name: "Política de Cookies",
+        link: "cookie-policy",
+      },
+    ],
+  },
+];
